requirements.txt
streamlit
anthropic
pycaret[full]
transformers
torch
datasets
huggingface_hub
pandas
scikit-learn
python-dotenv
requests
SpeechRecognition
pyaudio
streamlit-audio-recorder
fastapi
uvicorn
numpy
matplotlib
plotly